+++ cmdline.py
@@ -33,6 +33,7 @@ class CommonArguments:
     show_hint_match_dafny: bool = field(default=True, metadata={"help": "Show Dafny hint about pattern matching syntax"})
     coq_import_lia: bool = field(default=False, metadata={"help": "Import Lia for Coq"})
     n_trials: int = field(default=10, metadata={"help": "Number of trials to run on a specified experiment"})
+    clover: bool = field(default=False, metadata={"help": "Whether we are running clover experiments."})
     experiment_name: str = field(default="run.py", metadata={"help": "Pick an experiment to run "})
     mins_timeout: float = field(default=None, metadata={"help": "Set a default timeout for each trial "})
     seed: int = field(default=None, metadata={"help": "Set the seed for reproducible behavior"})
