+++ run_intermediate_expansion.py
@@ -14,19 +14,6 @@ from common_stats import stats
 
 import llm
 
-# TODO: this is not the right place for these guys
-min_lines = 5
-expansion_count = None
-
-# TODO: this is not the right place to define check_func
-def check_func(v):
-    lines = v.split('\n')  # Split the string into lines
-    for line in lines:
-        # Strip leading and trailing whitespace and check if it starts with '//'
-        if not line.lstrip().startswith('//'):
-            return True  # Found a line that doesn't start with '//'
-    return False  # All lines start with '//'
-
 def generate_complete(text, montecarlo, current_completion_depth=1):
     if current_completion_depth >= max_completion_depth:
         return None
